CSS set up to fit with theme of brandongioggia.com.
Added inline function to reload page to html banner.

The double alert error in case of a malformed link is proving more difficult to solve than I expected. It has something to do with the ajax calls happening in multiple threads I think. It requires more research, but is a small but, so I've added a line to the message and will move on with finishing for now.

All that remains for full functionality is adding the remaining passages so I have 20.